refactor- move HomePage to Pages file
