+++ src/Pages/HomePage/HomePage.module.scss
@@ -1,0 +1,43 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  font-family: sans-serif;
+  border: 1px solid #E6ECF0;
+  height: 100vh;
+  /* 用web-kit 改一下scrollbar樣式*/
+  overflow-y: scroll;
+}
+
+.page-title {
+  position: fixed;
+  width: 502.208px;
+  font-size: 24px;
+  line-height: 26px;
+  font-weight: 700;
+  padding: 24px;
+  background-color: var(--color-white);
+  opacity: 0.9;
+  cursor: pointer;
+
+  @media screen and (max-width:1200px) {
+    width: 398px;
+  }
+}
+
+/* 下推74px 讓上面首頁可以置頂*/
+.tweet-input {
+  margin-top: 74px;
+}
+
+/* 捲軸樣式調整 */
+::-webkit-scrollbar {
+  width: 10px;
+  background-color: #fafafa;
+  padding: 0 4px;
+}
+
+::-webkit-scrollbar-thumb {
+  background-color: #C1C1C1;
+  border-radius: 8px;
+  -webkit-box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
+}
